requirements.txt
appdirs==1.4.4
cx-Freeze==6.14.8
cx-Logging==3.1.0
lief==0.12.3
openvr==1.23.701
Pillow==9.5.0
pystray==0.19.4
python-osc==1.8.1
six==1.16.0
websockets==11.0.1
